global variables have been reimplemented
there may be some bugs which I missed because I didn't do extensive testing, but I'll weed them out over time.
The main changes to the system are:

-Global variables can no longer be created from the GUI, they must be created by defining a public attribute in an action index
-Global variables' values can be modified from the GUI, which will be written to the output action index, but there can be default values as well

I felt this change was necessary because in its previous state I felt there would be a lot of confusion around how global variables are supposed to be created and used (defined in file vs in GUI).  Now they're much less confusing and more effective as set-and-forget values, which is more or less their intention.